+++ bitv.ml
@@ -13,288 +13,106 @@
 (*                                                                        *)
 (**************************************************************************)
 
-(*s Bit vectors. The interface and part of the code are borrowed from the
-    [Array] module of the ocaml standard library (but things are simplified
-    here since we can always initialize a bit vector). This module also
-    provides bitwise operations. *)
+(* Bit vectors. The interface and part of the code are borrowed from the
+   [Array] module of the OCaml standard library (but things are simplified
+   here since we can always initialize a bit vector). This module also
+   provides bitwise operations. *)
 
-(*s We represent a bit vector by a vector of integers (field [bits]),
-    and we keep the information of the size of the bit vector since it
-    can not be found out with the size of the array (field [length]). *)
+(* We represent a bit vector by a string of bytes (field [bits]), and
+   we keep the information of the size of the bit vector (field
+   [length]) since it can not be figured out with the size of the
+   array. *)
+
+let[@inline] byte s i = Char.code (Bytes.unsafe_get s i)
+
+let set_byte s i x = Bytes.unsafe_set s i (Char.unsafe_chr x)
 
 type t = {
   length : int;
-  bits   : int array }
+  bits   : Bytes.t }
+  (* invariant: the unused high bits of the last byte are zeros (if any) *)
 
 let length v = v.length
 
-let[@inline] equal (v1:t) (v2:t) = v1 = v2
-(*s Perhaps the polymorphic equality is actually faster or good enough?
-    Did not test it. *)
+let[@inline] equal (v1: t) (v2: t) = v1 = v2
 
-(*s Each element of the array is an integer containing [bpi] bits, where
-    [bpi] is determined according to the machine word size. Since we do not
-    use the sign bit, [bpi] is 30 on a 32-bits machine and 62 on a 64-bits
-    machines. We maintain the following invariant:
-    {\em The unused bits of the last integer are always
-    zeros.} This is ensured by [create] and maintained in other functions
-    using [normalize]. [bit_j], [bit_not_j], [low_mask] and [up_mask]
-    are arrays used to extract and mask bits in a single integer. *)
-
-let bpi = Sys.word_size - 2
-
-let bit_j = Array.init bpi (fun j -> 1 lsl j)
-let bit_not_j = Array.init bpi (fun j -> max_int - bit_j.(j))
-
-let low_mask = Array.make (succ bpi) 0
-let _ =
-  for i = 1 to bpi do low_mask.(i) <- low_mask.(i-1) lor bit_j.(pred i) done
-
-let keep_lowest_bits a j = a land low_mask.(j)
-
-let high_mask = Array.init (succ bpi) (fun j -> low_mask.(j) lsl (bpi-j))
-
-let keep_highest_bits a j = a land high_mask.(j)
-
-let max_length = Sys.max_array_length * bpi
+let max_length = Sys.max_string_length * 8
 
 let exceeds_max_length n =
-  let s = n / bpi in
-  (if n mod bpi = 0 then s else s + 1) > Sys.max_array_length
+  let s = n / 8 in
+  (if n mod 8 = 0 then s else s + 1) > Sys.max_string_length
 
-(*s Creating and normalizing a bit vector is easy: it is just a matter of
-    taking care of the invariant. Copy is immediate. *)
+let low_mask = Array.init 9 (fun i -> (1 lsl i) - 1)
 
 let create n b =
-  if n < 0 || exceeds_max_length n then invalid_arg "Bitv.create";
-  let initv = if b then max_int else 0 in
-  let r = n mod bpi in
+  if n < 0 || n > max_length then invalid_arg "Bitv.create";
+  let initv = if b then 255 else 0 in
+  let q = n lsr 3 in
+  let r = n land 7 in
   if r = 0 then
-    { length = n; bits = Array.make (n / bpi) initv }
+    { length = n; bits = Bytes.make q (Char.chr initv) }
   else begin
-    let s = n / bpi in
-    let b = Array.make (succ s) initv in
-    b.(s) <- b.(s) land low_mask.(r);
-    { length = n; bits = b }
+    let s = Bytes.make (q + 1) (Char.chr initv) in
+    set_byte s q (initv land low_mask.(r));
+    { length = n; bits = s }
   end
 
 let normalize v =
-  let r = v.length mod bpi in
+  let r = v.length land 7 in
   if r > 0 then
     let b = v.bits in
-    let s = Array.length b in
-    b.(s-1) <- b.(s-1) land low_mask.(r)
+    let s = Bytes.length b in
+    set_byte b (s-1) ((byte b (s-1)) land low_mask.(r))
 
-let copy v = { length = v.length; bits = Array.copy v.bits }
-
-(*s Access and assignment. The [n]th bit of a bit vector is the [j]th
-    bit of the [i]th integer, where [i = n / bpi] and [j = n mod
-    bpi]. Both [i] and [j] and computed by the function [pos].
-    Accessing a bit is testing whether the result of the corresponding
-    mask operation is non-zero, and assigning it is done with a
-    bitwiwe operation: an {\em or} with [bit_j] to set it, and an {\em
-    and} with [bit_not_j] to unset it. *)
-
-let pos n = n / bpi, n mod bpi
+let copy v = { length = v.length; bits = Bytes.copy v.bits }
 
 let unsafe_get v n =
-  let (i,j) = pos n in
-  ((Array.unsafe_get v.bits i) land (Array.unsafe_get bit_j j)) > 0
-
-let unsafe_set v n b =
-  let (i,j) = pos n in
-  if b then
-    Array.unsafe_set v.bits i
-      ((Array.unsafe_get v.bits i) lor (Array.unsafe_get bit_j j))
-  else
-    Array.unsafe_set v.bits i
-      ((Array.unsafe_get v.bits i) land (Array.unsafe_get bit_not_j j))
-
-(*s The corresponding safe operations test the validiy of the access. *)
+  let i = n lsr 3 in
+  (byte v.bits i) land (1 lsl (n land 7)) > 0
 
 let get v n =
   if n < 0 || n >= v.length then invalid_arg "Bitv.get";
-  let (i,j) = pos n in
-  ((Array.unsafe_get v.bits i) land (Array.unsafe_get bit_j j)) > 0
+  unsafe_get v n
+
+let unsafe_set v n b =
+  let i = n lsr 3 in
+  let c = byte v.bits i in
+  let mask = 1 lsl (n land 7) in
+  set_byte v.bits i (if b then c lor mask else c land (lnot mask))
 
 let set v n b =
   if n < 0 || n >= v.length then invalid_arg "Bitv.set";
-  let (i,j) = pos n in
-  if b then
-    Array.unsafe_set v.bits i
-      ((Array.unsafe_get v.bits i) lor (Array.unsafe_get bit_j j))
-  else
-    Array.unsafe_set v.bits i
-      ((Array.unsafe_get v.bits i) land (Array.unsafe_get bit_not_j j))
+  unsafe_set v n b
 
-(*s [init] is implemented naively using [unsafe_set]. *)
+(* [init] is implemented naively using [unsafe_set]. *)
 
 let init n f =
   let v = create n false in
-  for i = 0 to pred n do
-    unsafe_set v i (f i)
-  done;
+  for i = 0 to pred n do unsafe_set v i (f i) done;
   v
-
-(*s Handling bits by packets is the key for efficiency of functions
-    [append], [concat], [sub] and [blit].
-    We start by a very general function [blit_bits a i m v n] which blits
-    the bits [i] to [i+m-1] of a native integer [a]
-    onto the bit vector [v] at index [n]. It assumes that [i..i+m-1] and
-    [n..n+m-1] are respectively valid subparts of [a] and [v].
-    It is optimized when the bits fit the lowest boundary of an integer
-    (case [j == 0]). *)
-
-let blit_bits a i m v n =
-  let (i',j) = pos n in
-  if j == 0 then
-    Array.unsafe_set v i'
-      ((keep_lowest_bits (a lsr i) m) lor
-       (keep_highest_bits (Array.unsafe_get v i') (bpi - m)))
-  else
-    let d = m + j - bpi in
-    if d > 0 then begin
-      Array.unsafe_set v i'
-	(((keep_lowest_bits (a lsr i) (bpi - j)) lsl j) lor
-	 (keep_lowest_bits (Array.unsafe_get v i') j));
-      Array.unsafe_set v (succ i')
-	((keep_lowest_bits (a lsr (i + bpi - j)) d) lor
-	 (keep_highest_bits (Array.unsafe_get v (succ i')) (bpi - d)))
-    end else
-      Array.unsafe_set v i'
-	(((keep_lowest_bits (a lsr i) m) lsl j) lor
-	 ((Array.unsafe_get v i') land (low_mask.(j) lor high_mask.(-d))))
-
-(*s [blit_int] implements [blit_bits] in the particular case when
-    [i=0] and [m=bpi] i.e. when we blit all the bits of [a]. *)
-
-let blit_int a v n =
-  let (i,j) = pos n in
-  if j == 0 then
-    Array.unsafe_set v i a
-  else begin
-    Array.unsafe_set v i
-      ( (keep_lowest_bits (Array.unsafe_get v i) j) lor
-       ((keep_lowest_bits a (bpi - j)) lsl j));
-    Array.unsafe_set v (succ i)
-      ((keep_highest_bits (Array.unsafe_get v (succ i)) (bpi - j)) lor
-       (a lsr (bpi - j)))
-  end
-
-(*s When blitting a subpart of a bit vector into another bit vector, there
-    are two possible cases: (1) all the bits are contained in a single integer
-    of the first bit vector, and a single call to [blit_bits] is the
-    only thing to do, or (2) the source bits overlap on several integers of
-    the source array, and then we do a loop of [blit_int], with two calls
-    to [blit_bits] for the two bounds. *)
-
-let unsafe_blit v1 ofs1 v2 ofs2 len =
-  if len > 0 then
-    let (bi,bj) = pos ofs1 in
-    let (ei,ej) = pos (ofs1 + len - 1) in
-    if bi == ei then
-      blit_bits (Array.unsafe_get v1 bi) bj len v2 ofs2
-    else begin
-      blit_bits (Array.unsafe_get v1 bi) bj (bpi - bj) v2 ofs2;
-      let n = ref (ofs2 + bpi - bj) in
-      for i = succ bi to pred ei do
-	blit_int (Array.unsafe_get v1 i) v2 !n;
-	n := !n + bpi
-      done;
-      blit_bits (Array.unsafe_get v1 ei) 0 (succ ej) v2 !n
-    end
-
-let blit v1 ofs1 v2 ofs2 len =
-  if len < 0 || ofs1 < 0 || ofs1 + len > v1.length
-             || ofs2 < 0 || ofs2 + len > v2.length
-  then invalid_arg "Bitv.blit";
-  unsafe_blit v1.bits ofs1 v2.bits ofs2 len
-
-(*s Extracting the subvector [ofs..ofs+len-1] of [v] is just creating a
-    new vector of length [len] and blitting the subvector of [v] inside. *)
-
-let sub v ofs len =
-  if ofs < 0 || len < 0 || ofs + len > v.length then invalid_arg "Bitv.sub";
-  let r = create len false in
-  unsafe_blit v.bits ofs r.bits 0 len;
-  r
-
-(*s The concatenation of two bit vectors [v1] and [v2] is obtained by
-    creating a vector for the result and blitting inside the two vectors.
-    [v1] is copied directly. *)
-
-let append v1 v2 =
-  let l1 = v1.length
-  and l2 = v2.length in
-  let r = create (l1 + l2) false in
-  let b1 = v1.bits in
-  let b2 = v2.bits in
-  let b = r.bits in
-  for i = 0 to Array.length b1 - 1 do
-    Array.unsafe_set b i (Array.unsafe_get b1 i)
-  done;
-  unsafe_blit b2 0 b l1 l2;
-  r
-
-(*s The concatenation of a list of bit vectors is obtained by iterating
-    [unsafe_blit]. *)
-
-let concat vl =
-  let size = List.fold_left (fun sz v -> sz + v.length) 0 vl in
-  let res = create size false in
-  let b = res.bits in
-  let pos = ref 0 in
-  List.iter
-    (fun v ->
-       let n = v.length in
-       unsafe_blit v.bits 0 b !pos n;
-       pos := !pos + n)
-    vl;
-  res
-
-(*s Filling is a particular case of blitting with a source made of all
-    ones || all zeros. Thus we instanciate [unsafe_blit], with 0 and
-    [max_int]. *)
-
-let blit_zeros v ofs len =
-  if len > 0 then
-    let (bi,bj) = pos ofs in
-    let (ei,ej) = pos (ofs + len - 1) in
-    if bi == ei then
-      blit_bits 0 bj len v ofs
-    else begin
-      blit_bits 0 bj (bpi - bj) v ofs;
-      let n = ref (ofs + bpi - bj) in
-      for _i = succ bi to pred ei do
-	blit_int 0 v !n;
-	n := !n + bpi
-      done;
-      blit_bits 0 0 (succ ej) v !n
-    end
-
-let blit_ones v ofs len =
-  if len > 0 then
-    let (bi,bj) = pos ofs in
-    let (ei,ej) = pos (ofs + len - 1) in
-    if bi == ei then
-      blit_bits max_int bj len v ofs
-    else begin
-      blit_bits max_int bj (bpi - bj) v ofs;
-      let n = ref (ofs + bpi - bj) in
-      for _i = succ bi to pred ei do
-	blit_int max_int v !n;
-	n := !n + bpi
-      done;
-      blit_bits max_int 0 (succ ej) v !n
-    end
 
 let fill v ofs len b =
   if ofs < 0 || len < 0 || ofs + len > v.length then invalid_arg "Bitv.fill";
-  if b then blit_ones v.bits ofs len else blit_zeros v.bits ofs len
+  if len > 0 then (
+  (* incomplete first byte, if any (8-r bits) *)
+  let r = ofs land 7 in
+  let first = if r = 0 then 0 else min len (8-r) in
+  for i = ofs to ofs + first-1 do unsafe_set v i b done;
+  (* full bytes in the middle *)
+  let start = (ofs + first) lsr 3 in
+  let n = (len - first) lsr 3 in
+  let x = Char.chr (if b then 0xFF else 0) in
+  for i = start to start + n - 1 do Bytes.unsafe_set v.bits i x done;
+  (* incomplete last byte, if any *)
+  let s = (len - first) land 7 in
+  if s > 0 then
+    let stop = ofs + len in
+    for i = stop - s to stop - 1 do unsafe_set v i b done
+  )
 
-(*s All the iterators are implemented as for traditional arrays, using
-    [unsafe_get]. For [iter] and [map], we do not precompute [(f
-    true)] and [(f false)] since [f] may have side-effects. *)
+(* All the iterators are implemented as for traditional arrays, using
+   [unsafe_get]. For [iter] and [map], we do not precompute [(f
+   true)] and [(f false)] since [f] may have side-effects. *)
 
 let iter f v =
   for i = 0 to v.length - 1 do f (unsafe_get v i) done
@@ -346,133 +164,215 @@ let foldi_right f v x =
   done;
   !r
 
-(*s Population count *)
+(*
+let iteri_true f v =
+  Bytes.iteri
+    (fun i x -> let x = Char.code x in if x != 0 then begin
+      let i = i lsl 3 in
+      for j = 0 to 7 do if x land (1 lsl j) > 0 then f (i + j) done
+    end)
+    v.bits
+*)
+(* slightly more efficient by precomputing NTZ *)
+let ntz = Array.make 256 0
+let () = for i = 0 to 7 do ntz.(1 lsl i) <- i done
+let ntz8 x = Array.unsafe_get ntz x
+
+let iteri_true f v =
+  Bytes.iteri
+    (fun i c ->
+       let i_bpi = i lsl 3 in
+       let rec visit x =
+	 if x != 0 then begin
+	   let b = x land (-x) in
+	   f (i_bpi + ntz8 b);
+	   visit (x - b)
+	 end
+       in
+       visit (Char.code c))
+    v.bits
+
+(* Population count *)
 
 let rec naive_pop x =
-  assert (x < 0x10000);
+  assert (x < 0x100);
   if x = 0 then 0 else 1 + naive_pop (x - (x land -x))
 
-let pop16 = Array.init 0x10000 naive_pop
-let pop16 n = Array.unsafe_get pop16 n
-
-let popi x = match Sys.word_size with
-  | 32 -> pop16 (x land 0xffff) + pop16 ((x lsr 16) land 0xffff)
-  | 64 -> pop16 (x land 0xffff) + pop16 ((x lsr 16) land 0xffff)
-        + pop16 ((x lsr 32) land 0xffff) + pop16 ((x lsr 48) land 0xffff)
-  | _ -> assert false
+let pop8 = Array.init 0x100 naive_pop
+let pop8 n = Array.unsafe_get pop8 n
 
 let pop v =
-  Array.fold_left (fun acc n -> acc + popi n) 0 v.bits
+  let n = Bytes.length v.bits in
+  let b = v.bits in
+  let rec loop acc i =
+    if i >= n then acc else loop (acc + pop8 (byte b i)) (i + 1) in
+  loop 0 0
 
-(*s Number of trailing zeros (on a 32-bit machine) *)
-
-let hash32 x = ((0x34ca8b09 * x) land 0x3fffffff) lsr 24
-let ntz_arr32 = Array.make 64 0
-let () = for i = 0 to 30 do ntz_arr32.(hash32 (1 lsl i)) <- i done
-let ntz32 x = if x == 0 then 31 else ntz_arr32.(hash32 (x land (-x)))
-
-let iteri_true_ntz32 f v =
-  Array.iteri
-    (fun i n ->
-       let i_bpi = i * bpi in
-       let rec visit x =
-	 if x != 0 then begin
-	   let b = x land (-x) in
-	   f (i_bpi + ntz32 b);
-	   visit (x - b)
-	 end
-       in
-       visit n)
-    v.bits
-
-let martin_constant = (0x03f79d71b lsl 28) lor 0x4ca8b09 (*0x03f79d71b4ca8b09*)
-let hash64 x = ((martin_constant * x) land max_int) lsr 56
-let ntz_arr64 = Array.make 64 0
-let () = if Sys.word_size >= 64 then
-  for i = 0 to 62 do ntz_arr64.(hash64 (1 lsl i)) <- i done
-let ntz64 x = if x == 0 then 63 else ntz_arr64.(hash64 (x land (-x)))
-
-let iteri_true_ntz64 f v =
-  Array.iteri
-    (fun i n ->
-       let i_bpi = i * bpi in
-       let rec visit x =
-	 if x != 0 then begin
-	   let b = x land (-x) in
-	   f (i_bpi + ntz64 b);
-	   visit (x - b)
-	 end
-       in
-       visit n)
-    v.bits
-
-let iteri_true = match Sys.word_size with
-  | 32 -> iteri_true_ntz32
-  | 64 -> iteri_true_ntz64
-  | _ -> assert false
-
-(*s Bitwise operations. It is straigthforward, since bitwise operations
-    can be realized by the corresponding bitwise operations over integers.
-    However, one has to take care of normalizing the result of [bwnot]
-    which introduces ones in highest significant positions. *)
+(* Bitwise operations. It is straigthforward, since bitwise operations
+   can be realized by the corresponding bitwise operations over integers.
+   However, one has to take care of normalizing the result of [bwnot]
+   which introduces ones in highest significant positions. *)
 
 let bw_and v1 v2 =
   let l = v1.length in
   if l <> v2.length then invalid_arg "Bitv.bw_and";
   let b1 = v1.bits
   and b2 = v2.bits in
-  let n = Array.length b1 in
-  let a = Array.make n 0 in
+  let n = Bytes.length b1 in
+  let a = Bytes.make n (Char.chr 0) in
   for i = 0 to n - 1 do
-    a.(i) <- b1.(i) land b2.(i)
+    set_byte a i ((byte b1 i) land (byte b2 i))
   done;
   { length = l; bits = a }
-
-let tanimoto v1 v2 =
-  let l = v1.length in
-  if l <> v2.length then invalid_arg "Bitv.tanimoto";
-  let a = pop v1 in
-  let b = pop v2 in
-  let c = pop (bw_and v1 v2) in
-  (float c) /. (float (a + b - c))
 
 let bw_or v1 v2 =
   let l = v1.length in
   if l <> v2.length then invalid_arg "Bitv.bw_or";
   let b1 = v1.bits
   and b2 = v2.bits in
-  let n = Array.length b1 in
-  let a = Array.make n 0 in
+  let n = Bytes.length b1 in
+  let a = Bytes.make n (Char.chr 0) in
   for i = 0 to n - 1 do
-    a.(i) <- b1.(i) lor b2.(i)
+    set_byte a i ((byte b1 i) lor (byte b2 i))
   done;
   { length = l; bits = a }
 
 let bw_xor v1 v2 =
   let l = v1.length in
-  if l <> v2.length then invalid_arg "Bitv.bw_xor";
+  if l <> v2.length then invalid_arg "Bitv.bw_or";
   let b1 = v1.bits
   and b2 = v2.bits in
-  let n = Array.length b1 in
-  let a = Array.make n 0 in
+  let n = Bytes.length b1 in
+  let a = Bytes.make n (Char.chr 0) in
   for i = 0 to n - 1 do
-    a.(i) <- b1.(i) lxor b2.(i)
+    set_byte a i ((byte b1 i) lxor (byte b2 i))
   done;
   { length = l; bits = a }
 
 let bw_not v =
   let b = v.bits in
-  let n = Array.length b in
-  let a = Array.make n 0 in
+  let n = Bytes.length b in
+  let a = Bytes.make n (Char.chr 0) in
   for i = 0 to n - 1 do
-    a.(i) <- max_int land (lnot b.(i))
+    set_byte a i (255 land (lnot (byte b i)))
   done;
   let r = { length = v.length; bits = a } in
   normalize r;
   r
 
-(*s Shift operations. It is easy to reuse [unsafe_blit], although it is
-    probably slightly less efficient than a ad-hoc piece of code. *)
+(* Coercions to/from lists of integers *)
+
+let of_list l =
+  let n = 1 + List.fold_left max (-1) l in
+  if n < 0 || n > max_length then invalid_arg "Bitv.of_list";
+  let b = create n false in
+  let add_element i =
+    (* negative numbers are invalid *)
+    if i < 0 then invalid_arg "Bitv.of_list";
+    unsafe_set b i true
+  in
+  List.iter add_element l;
+  b
+
+let of_list_with_length l len =
+  if len < 0 || len > max_length then invalid_arg "Bitv.of_list_with_length";
+  let b = create len false in
+  let add_element i =
+    if i < 0 || i >= len then invalid_arg "Bitv.of_list_with_length";
+    unsafe_set b i true
+  in
+  List.iter add_element l;
+  b
+
+let to_list b =
+  let n = length b in
+  let rec make i acc =
+    if i < 0 then acc
+    else make (pred i) (if unsafe_get b i then i :: acc else acc)
+  in
+  make (pred n) []
+
+let[@inline] pos n = n lsr 3, n land 7
+
+let unsafe_getb b n =
+  let i = n lsr 3 in
+  (byte b i) land (1 lsl (n land 7)) > 0
+
+let unsafe_setb b n v =
+  let i = n lsr 3 in
+  let c = byte b i in
+  let mask = 1 lsl (n land 7) in
+  set_byte b i (if v then c lor mask else c land (lnot mask))
+
+(* Copies v1[ofs1..ofs1+len[ into v2[ofs2..ofs2+len[ *)
+let unsafe_blit b1 ofs1 b2 ofs2 len =
+  if len > 0 then
+    if ofs1 land 7 = 0 && ofs2 land 7 = 0 && len land 7 = 0 then
+      Bytes.blit b1 (ofs1 lsr 3) b2 (ofs2 lsr 3) (len lsr 3)
+    else
+      for i = 0 to len - 1 do
+        unsafe_setb b2 (ofs2 + i) (unsafe_getb b1 (ofs1 + i))
+      done
+      (* TODO: improve in other cases when bytes can be batch-copied *)
+
+let blit v1 ofs1 v2 ofs2 len =
+  if len < 0 || ofs1 < 0 || ofs1 + len > v1.length
+             || ofs2 < 0 || ofs2 + len > v2.length
+  then invalid_arg "Bitv.blit";
+  unsafe_blit v1.bits ofs1 v2.bits ofs2 len
+
+let sub v ofs len =
+  if ofs < 0 || len < 0 || ofs + len > v.length then invalid_arg "Bitv.sub";
+  let r = create len false in
+  unsafe_blit v.bits ofs r.bits 0 len;
+  r
+
+let append v1 v2 =
+  let l1 = v1.length
+  and l2 = v2.length in
+  let r = create (l1 + l2) false in
+  let b1 = v1.bits in
+  let b = r.bits in
+  Bytes.blit b1 0 b 0 (Bytes.length b1);
+  unsafe_blit v2.bits 0 b l1 l2;
+  r
+
+let concat vl =
+  let size = List.fold_left (fun sz v -> sz + v.length) 0 vl in
+  let res = create size false in
+  let b = res.bits in
+  let pos = ref 0 in
+  List.iter
+    (fun v ->
+       let n = v.length in
+       unsafe_blit v.bits 0 b !pos n;
+       pos := !pos + n)
+    vl;
+  res
+
+(* Testing for all zeros and all ones. *)
+
+let all_zeros v =
+  let b = v.bits in
+  let n = Bytes.length b in
+  let rec test i = i == n || (byte b i == 0) && test (succ i) in
+  test 0
+
+let all_ones v =
+  let b = v.bits in
+  let n = Bytes.length b in
+  let rec test i =
+    if i == n - 1 then
+      let m = v.length land 7 in
+      byte b i == if m == 0 then 0xFF else low_mask.(m)
+    else
+      byte b i == 0xFF && test (succ i)
+  in
+  n = 0 || test 0
+
+
+(* Shift operations. It is easy to reuse [unsafe_blit], although it is
+   probably slightly less efficient than a ad-hoc piece of code. *)
 
 let rec shiftl v d =
   if d == 0 then
@@ -498,8 +398,8 @@ and shiftr v d =
     r
   end
 
-(*s Rotate operations. It is easy to reuse [unsafe_blit], although it is
-    probably slightly less efficient than an ad-hoc piece of code. *)
+(* Rotate operations. It is easy to reuse [unsafe_blit], although it is
+   probably slightly less efficient than an ad-hoc piece of code. *)
 
 let rec rotatel v d =
   if d < 0 then
@@ -531,29 +431,28 @@ and rotater v d =
     r
   end
 
-(*s Testing for all zeros and all ones. *)
+(* Iteration on all bit vectors of length [n] using a Gray code. *)
 
-let all_zeros v =
-  let b = v.bits in
-  let n = Array.length b in
-  let rec test i =
-    (i == n) || ((Array.unsafe_get b i == 0) && test (succ i))
+let first_set v n =
+  let rec lookup i =
+    if i = n then raise Not_found ;
+    if unsafe_get v i then i else lookup (i + 1)
   in
-  test 0
+  lookup 0
 
-let all_ones v =
-  let b = v.bits in
-  let n = Array.length b in
-  let rec test i =
-    if i == n - 1 then
-      let m = v.length mod bpi in
-      (Array.unsafe_get b i) == (if m == 0 then max_int else low_mask.(m))
-    else
-      ((Array.unsafe_get b i) == max_int) && test (succ i)
+let gray_iter f n =
+  let bv = create n false in
+  let rec iter () =
+    f bv;
+    unsafe_set bv 0 (not (unsafe_get bv 0));
+    f bv;
+    let pos = succ (first_set bv n) in
+    if pos < n then begin
+      unsafe_set bv pos (not (unsafe_get bv pos));
+      iter ()
+    end
   in
-  test 0
-
-(*s Conversions to and from strings. *)
+  if n > 0 then iter ()
 
 module S(I : sig val least_first : bool end) = struct
 
@@ -583,7 +482,15 @@ end
 module L = S(struct let least_first = true end)
 module M = S(struct let least_first = false end)
 
-(*s Input/output in a machine-independent format. *)
+let tanimoto v1 v2 =
+  let l = v1.length in
+  if l <> v2.length then invalid_arg "Bitv.tanimoto";
+  let a = pop v1 in
+  let b = pop v2 in
+  let c = pop (bw_and v1 v2) in
+  (float c) /. (float (a + b - c))
+
+(* Input/output in a machine-independent format. *)
 
 let bytes_of_int x =
   Bytes.init 8 (fun i -> Char.chr ((x lsr (8 * i)) land 0xFF))
@@ -596,216 +503,95 @@ let int_of_bytes b =
   in
   build 0 7
 
+let nb_of_bytes len =
+  len lsr 3 + if len land 7 = 0 then 0 else 1
+
 let to_bin write v =
-  let len = length v in
-  let rec loop i pow byte =
-    let byte = if unsafe_get v i then byte lor pow else byte in
-    if i = len - 1 then
-      write byte
-    else if i mod 8 = 7 then begin
-      write byte;
-      loop (i + 1) 1 0
-    end else
-      loop (i + 1) (pow * 2) byte
-  in
-  bytes_of_int len
-  |> Bytes.iter (fun b -> Char.code b |> write);
-  if len > 0 then loop 0 1 0
+  bytes_of_int v.length |> Bytes.iter write;
+  Bytes.iter write v.bits
 
 let output_bin out_ch v =
-  let write = output_byte out_ch in
+  let write = output_char out_ch in
   to_bin write v
 
 let to_bytes t =
   let buf = Buffer.create 0 in
-  let write i = Buffer.add_char buf (Char.chr i) in
+  let write x = Buffer.add_char buf x in
   to_bin write t;
   Buffer.to_bytes buf
 
 let of_bin read =
-  let len = Bytes.init 8 (fun _ -> read () |> Char.chr) |> int_of_bytes in
-  let bits = create len false in
-  let rec loop i byte =
-    if i < len then begin
-      let byte = if i mod 8 = 0 then read () else byte in
-      if byte land 1 = 1 then unsafe_set bits i true;
-      loop (i+1) (byte / 2)
-    end
-  in
-  if len > 0 then loop 0 0;
-  bits
+  let len = Bytes.init 8 (fun _ -> read ()) |> int_of_bytes in
+  let v = create len false in
+  let b = v.bits in
+  for i = 0 to Bytes.length b - 1 do
+    Bytes.unsafe_set b i (read ())
+  done;
+  v
 
 let input_bin in_ch =
-  let read () = input_byte in_ch in
+  let read () = input_char in_ch in
   of_bin read
 
 let of_bytes b =
   let read =
     let p = ref 0 in
-    fun () ->
-      let ret = Bytes.get b !p |> Char.code in
-      incr p;
-      ret
-  in
+    fun () -> let ret = Bytes.get b !p in incr p; ret in
   of_bin read
-
-(* Iteration on all bit vectors of length [n] using a Gray code. *)
-
-let first_set v n =
-  let rec lookup i =
-    if i = n then raise Not_found ;
-    if unsafe_get v i then i else lookup (i + 1)
-  in
-  lookup 0
-
-let gray_iter f n =
-  let bv = create n false in
-  let rec iter () =
-    f bv;
-    unsafe_set bv 0 (not (unsafe_get bv 0));
-    f bv;
-    let pos = succ (first_set bv n) in
-    if pos < n then begin
-      unsafe_set bv pos (not (unsafe_get bv pos));
-      iter ()
-    end
-  in
-  if n > 0 then iter ()
-
-
-(*s Coercions to/from lists of integers *)
-
-let of_list l =
-  let n = List.fold_left max 0 l in
-  let b = create (succ n) false in
-  let add_element i =
-    (* negative numbers are invalid *)
-    if i < 0 then invalid_arg "Bitv.of_list";
-    unsafe_set b i true
-  in
-  List.iter add_element l;
-  b
-
-let of_list_with_length l len =
-  let b = create len false in
-  let add_element i =
-    if i < 0 || i >= len then invalid_arg "Bitv.of_list_with_length";
-    unsafe_set b i true
-  in
-  List.iter add_element l;
-  b
-
-let to_list b =
-  let n = length b in
-  let rec make i acc =
-    if i < 0 then acc
-    else make (pred i) (if unsafe_get b i then i :: acc else acc)
-  in
-  make (pred n) []
-
 
 (*s To/from integers. *)
 
-(* [int] *)
-let of_int_us i =
-  { length = bpi; bits = [| i land max_int |] }
+let of_int_gen len getbyte getbit =
+  let v = create len false in
+  for i = 0 to (len lsr 3) - 1 do set_byte v.bits i (getbyte i) done;
+  for i = len land (lnot 7) to len - 1 do unsafe_set v i (getbit i) done;
+  v
+
+let getbyte x i = (x lsr (8*i)) land 0xFF
+let getbit x i = (x lsr i) land 1 > 0
+let of_int_us x =
+  of_int_gen (Sys.int_size - 1) (getbyte x) (getbit x)
+let of_int_s x =
+  of_int_gen Sys.int_size (getbyte x) (getbit x)
+
+let to_int_gen zero shiftor v =
+  let x = ref zero in
+  Bytes.iteri (fun i c -> x := shiftor !x (Char.code c) (8*i)) v.bits;
+  !x
+
+let shiftor x b i = x lor (b lsl i)
 let to_int_us v =
-  if v.length < bpi then invalid_arg "Bitv.to_int_us";
-  v.bits.(0)
-
-let of_int_s i =
-  { length = succ bpi; bits = [| i land max_int; (i lsr bpi) land 1 |] }
+  (* if v.length < Sys.int_size - 1 then invalid_arg "Bitv.to_int_us"; *)
+  to_int_gen 0 shiftor v
 let to_int_s v =
-  if v.length < succ bpi then invalid_arg "Bitv.to_int_s";
-  v.bits.(0) lor (v.bits.(1) lsl bpi)
+  (* if v.length < Sys.int_size then invalid_arg "Bitv.to_int_s"; *)
+  to_int_gen 0 shiftor v
 
-(* [Int32] *)
-let of_int32_us i = match Sys.word_size with
-  | 32 -> { length = 31;
-	    bits = [| (Int32.to_int i) land max_int;
-		      let hi = Int32.shift_right_logical i 30 in
-		      (Int32.to_int hi) land 1 |] }
-  | 64 -> { length = 31; bits = [| (Int32.to_int i) land 0x7fffffff |] }
-  | _ -> assert false
-let to_int32_us v =
-  if v.length < 31 then invalid_arg "Bitv.to_int32_us";
-  match Sys.word_size with
-    | 32 ->
-	Int32.logor (Int32.of_int v.bits.(0))
-	            (Int32.shift_left (Int32.of_int (v.bits.(1) land 1)) 30)
-    | 64 ->
-	Int32.of_int (v.bits.(0) land 0x7fffffff)
-    | _ -> assert false
+let getbyte32 x i =
+  Int32.to_int (Int32.logand (Int32.shift_right x (8*i)) 0xFFl)
+let getbit32 x i =
+  Int32.logand (Int32.shift_right x i) 1l > 0l
+let of_int32_us x =
+  of_int_gen 31 (getbyte32 x) (getbit32 x)
+let of_int32_s x =
+  of_int_gen 32 (getbyte32 x) (getbit32 x)
 
-(* this is 0xffffffff (ocaml >= 3.08 checks for literal overflow) *)
-let ffffffff = (0xffff lsl 16) lor 0xffff
+let shiftor32 x b i = Int32.logor x (Int32.shift_left (Int32.of_int b) i)
+let to_int32_us v = to_int_gen 0l shiftor32 v
+let to_int32_s v = to_int_gen 0l shiftor32 v
 
-let of_int32_s i = match Sys.word_size with
-  | 32 -> { length = 32;
-	    bits = [| (Int32.to_int i) land max_int;
-		      let hi = Int32.shift_right_logical i 30 in
-		      (Int32.to_int hi) land 3 |] }
-  | 64 -> { length = 32; bits = [| (Int32.to_int i) land ffffffff |] }
-  | _ -> assert false
-let to_int32_s v =
-  if v.length < 32 then invalid_arg "Bitv.to_int32_s";
-  match Sys.word_size with
-    | 32 ->
-	Int32.logor (Int32.of_int v.bits.(0))
-	            (Int32.shift_left (Int32.of_int (v.bits.(1) land 3)) 30)
-    | 64 ->
-	Int32.of_int (v.bits.(0) land ffffffff)
-    | _ -> assert false
+let getbyte64 x i =
+  Int64.to_int (Int64.logand (Int64.shift_right x (8*i)) 0xFFL)
+let getbit64 x i =
+  Int64.logand (Int64.shift_right x i) 1L > 0L
+let of_int64_us x =
+  of_int_gen 63 (getbyte64 x) (getbit64 x)
+let of_int64_s x =
+  of_int_gen 64 (getbyte64 x) (getbit64 x)
 
-(* [Int64] *)
-let of_int64_us i = match Sys.word_size with
-  | 32 -> { length = 63;
-	    bits = [| (Int64.to_int i) land max_int;
-		      (let mi = Int64.shift_right_logical i 30 in
-		       (Int64.to_int mi) land max_int);
-		      let hi = Int64.shift_right_logical i 60 in
-		      (Int64.to_int hi) land 7 |] }
-  | 64 -> { length = 63;
-	    bits = [| (Int64.to_int i) land max_int;
-		      let hi = Int64.shift_right_logical i 62 in
-		      (Int64.to_int hi) land 1 |] }
-  | _ -> assert false
-let to_int64_us v =
-  if v.length < 63 then invalid_arg "Bitv.to_int64_us";
-  match Sys.word_size with
-    | 32 ->
-	Int64.logor (Int64.of_int v.bits.(0))
-        (Int64.logor (Int64.shift_left (Int64.of_int v.bits.(1)) 30)
-                     (Int64.shift_left (Int64.of_int (v.bits.(2) land 7)) 60))
-    | 64 ->
-	Int64.logor (Int64.of_int v.bits.(0))
-                    (Int64.shift_left (Int64.of_int (v.bits.(1) land 1)) 62)
-    | _ ->
-        assert false
-
-let of_int64_s i = match Sys.word_size with
-  | 32 -> { length = 64;
-	    bits = [| (Int64.to_int i) land max_int;
-		      (let mi = Int64.shift_right_logical i 30 in
-		       (Int64.to_int mi) land max_int);
-		      let hi = Int64.shift_right_logical i 60 in
-		      (Int64.to_int hi) land 15 |] }
-  | 64 -> { length = 64;
-            bits = [| (Int64.to_int i) land max_int;
-                      let hi = Int64.shift_right_logical i 62 in
-		      (Int64.to_int hi) land 3 |] }
-  | _ -> assert false
-let to_int64_s v =
-  if v.length < 64 then invalid_arg "Bitv.to_int64_s";
-  match Sys.word_size with
-    | 32 ->
-	Int64.logor (Int64.of_int v.bits.(0))
-        (Int64.logor (Int64.shift_left (Int64.of_int v.bits.(1)) 30)
-                     (Int64.shift_left (Int64.of_int (v.bits.(2) land 15)) 60))
-    | 64 ->
-	Int64.logor (Int64.of_int v.bits.(0))
-                    (Int64.shift_left (Int64.of_int (v.bits.(1) land 3)) 62)
-    | _ -> assert false
+let shiftor64 x b i = Int64.logor x (Int64.shift_left (Int64.of_int b) i)
+let to_int64_us v = to_int_gen 0L shiftor64 v
+let to_int64_s v = to_int_gen 0L shiftor64 v
 
 (* [Nativeint] *)
 let select_of f32 f64 = match Sys.word_size with