+++ bitv.mli
@@ -218,7 +218,7 @@ val of_bytes: bytes -> t
 
 (** {2 Conversions to and from lists of integers}
 
-    The list gives the indices of bits which are set (ie [true]). *)
+    The list gives the indices of bits which are set (i.e. [true]). *)
 
 val to_list : t -> int list
 val of_list : int list -> t
@@ -226,10 +226,8 @@ val of_list_with_length : int list -> int -> t
 
 (** {2 Interpretation of bit vectors as integers}
 
-    Least significant bit
-    comes first (ie is at index 0 in the bit vector).
-    [to_xxx] functions truncate when the bit vector is too wide,
-    and raise [Invalid_argument] when it is too short.
+    Least significant bit comes first (ie is at index 0 in the bit vector).
+    [to_xxx] functions truncate when the bit vector is too wide.
     Suffix [_s] means that sign bit is kept,
     and [_us] that it is discarded. *)